feat: Add note creation functionality using noteService
This commit adds the functionality to create a new note using the noteService module. The `NewNote` component now makes an asynchronous request to the server to create a new note with the provided content. The response from the server is then dispatched to the Redux store using the `createNote` action.